fix(*-observation): allow binding extended Map/Set

+++ src/map-observation.js
@@ -16,28 +16,40 @@ class ModifyMapObserver extends ModifyCollectionObserver {
   static create(taskQueue, map) {
     let observer = new ModifyMapObserver(taskQueue, map);
 
+    let methods = ['set', 'delete', 'clear'];
+    if (methods.find(m => mapProto[m] !== map[m])) {
+      mapProto = {};
+      methods.forEach(m => mapProto[m] = map[m]);
+    }
+
     map['set'] = function () {
+      let hasValue = map.has(arguments[0]);
+      let type = hasValue ? 'update' : 'add';
       let oldValue = map.get(arguments[0]);
-      let type = typeof oldValue !== 'undefined' ? 'update' : 'add';
       let methodCallResult = mapProto['set'].apply(map, arguments);
-      observer.addChangeRecord({
-        type: type,
-        object: map,
-        key: arguments[0],
-        oldValue: oldValue
-      });
+      if (!hasValue || oldValue !== map.get(arguments[0])) {
+        observer.addChangeRecord({
+          type: type,
+          object: map,
+          key: arguments[0],
+          oldValue: oldValue
+        });
+      }
       return methodCallResult;
     }
 
     map['delete'] = function () {
+      let hasValue = map.has(arguments[0]);
       let oldValue = map.get(arguments[0]);
       let methodCallResult = mapProto['delete'].apply(map, arguments);
-      observer.addChangeRecord({
-        type: 'delete',
-        object: map,
-        key: arguments[0],
-        oldValue: oldValue
-      });
+      if (hasValue) {
+        observer.addChangeRecord({
+          type: 'delete',
+          object: map,
+          key: arguments[0],
+          oldValue: oldValue
+        });
+      }
       return methodCallResult;
     }
